+++ CHANGES.md
@@ -1,3 +1,9 @@
+# Version 1.3.0
+- Deprecate `AndroidFs::create_file_writable`.
+- Add `AndroidFs::create_file`.
+- Add `PrivateStorage::create_file`.
+- Update documents.
+
 # Version 1.2.0
 - Add `AndroidFs::get_mime_type`.
 - Update documents.